+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortTool.java
@@ -2,20 +2,15 @@ package ai.mcpdirect.backend.dao.entity.aitool;
 
 import ai.mcpdirect.backend.util.ID;
 
-public class AIPortTool {
-    public long id;
-    public long makerId;
+public class AIPortTool extends AIPortToolStub{
     public int status;
     public long lastUpdated;
-    public String name;
     public int hash;
     public String metaData;
     public String tags;
     public long agentId;
     public int agentStatus;
     public int makerStatus;
-    public int usage;
-    public long userId;
     public AIPortTool() {}
 
     public AIPortTool(long makerId, int status, long lastUpdated, String name, int hash, String metaData, String tags) {