+++ src/main/java/ai/mcpdirect/backend/dao/entity/aitool/AIPortToolStub.java
@@ -1,0 +1,12 @@
+package ai.mcpdirect.backend.dao.entity.aitool;
+
+import ai.mcpdirect.backend.util.ID;
+
+public class AIPortToolStub {
+    public long id;
+    public long makerId;
+    public long removed;
+    public String name;
+    public int usage;
+    public long userId;
+}