+++ src/main/java/ai/mcpdirect/backend/dao/mapper/aitool/ToolAccessKeyMapper.java
@@ -1,21 +1,19 @@
-package ai.mcpdirect.backend.dao.mapper.account;
+package ai.mcpdirect.backend.dao.mapper.aitool;
 
-import ai.mcpdirect.backend.dao.entity.account.AIPortAccessKey;
-
-import ai.mcpdirect.backend.dao.entity.account.AIPortAccessKeyCredential;
+import ai.mcpdirect.backend.dao.entity.aitool.AIPortToolAccessKey;
+import ai.mcpdirect.backend.dao.entity.aitool.AIPortToolAccessKeyCredential;
 import org.apache.ibatis.annotations.*;
 
 import java.util.List;
 
-public interface AccessKeyMapper {
-    String accessKeyTable="account.access_key";
+public interface ToolAccessKeyMapper {
+    String accessKeyTable="aitool.tool_access_key";
     String selectAccessKey = """
             SELECT id,
             status,"name",
             effective_date effectiveDate,
             expiration_date expirationDate,
             user_id userId,
-            user_roles userRoles,
             usage usage,
             created FROM
             """+accessKeyTable+"\n";
@@ -26,7 +24,6 @@ public interface AccessKeyMapper {
             effective_date effectiveDate,
             expiration_date expirationDate,
             user_id userId,
-            user_roles userRoles,
             usage usage,
             created FROM
             """+accessKeyTable+"\n";
@@ -36,36 +33,36 @@ public interface AccessKeyMapper {
 //    AIPortAccessKeyCredential selectAccessKeyCredentialById(@Param("userId")long userId,@Param("accessKeyId") int accessKeyId);
 
     @Select(selectAccessKeyCredential+"where id=#{accessKeyId}")
-    AIPortAccessKeyCredential selectAccessKeyCredentialById(@Param("accessKeyId") long accessKeyId);
+    AIPortToolAccessKeyCredential selectAccessKeyCredentialById(@Param("accessKeyId") long accessKeyId);
 
 
     @Select(selectAccessKey+"where secret_key=#{key}")
-    AIPortAccessKey selectAccessKey(String key);
+    AIPortToolAccessKey selectAccessKey(String key);
     @Select(selectAccessKey+" WHERE user_Id=#{userId} and id=#{id}")
-    AIPortAccessKey selectAccessKeyById(@Param("userId")long userId,@Param("id") long id);
+    AIPortToolAccessKey selectAccessKeyById(@Param("userId")long userId,@Param("id") long id);
 
     @Select("<script>"+selectAccessKey+" WHERE user_Id=#{userId} AND id IN\n"+"""
                                 <foreach item='item' index='index' collection='idList' open='(' separator=', ' close=')'>
                                 #{item}
                                 </foreach></script>""")
-    List<AIPortAccessKey> selectAccessKeyByIds(@Param("userId")long userId,@Param("idList") List<Integer> idList);
+    List<AIPortToolAccessKey> selectAccessKeyByIds(@Param("userId")long userId, @Param("idList") List<Integer> idList);
 
     @Select(selectAccessKey+"where user_id=#{userId}")
-    List<AIPortAccessKey> selectAccessKeyByUserId(long userId);
+    List<AIPortToolAccessKey> selectAccessKeyByUserId(long userId);
 
     @Insert("INSERT INTO "+accessKeyTable+
             "(id, secret_key,\"name\",status,effective_date, expiration_date, user_id, user_roles,created)VALUES" +
             "(#{id},#{secretKey},#{name},#{status},#{effectiveDate},#{expirationDate},#{userId},#{userRoles},#{created})")
-    void insertAccessKeyCredential(AIPortAccessKeyCredential accessKey);
+    void insertAccessKeyCredential(AIPortToolAccessKeyCredential accessKey);
 
     @Update("UPDATE "+accessKeyTable+" SET name=#{name}, status=#{status} WHERE user_Id=#{userId} and id=#{id}>")
-    int updateAccessKey(AIPortAccessKey accessKey);
+    int updateAccessKey(AIPortToolAccessKey accessKey);
 
     @Update("UPDATE "+accessKeyTable+" SET name=#{name} WHERE user_Id=#{userId} and id=#{id}")
-    int updateAccessKeyName(AIPortAccessKey accessKey);
+    int updateAccessKeyName(AIPortToolAccessKey accessKey);
 
     @Update("UPDATE "+accessKeyTable+" SET status=#{status} WHERE user_Id=#{userId} and id=#{id}")
-    int updateAccessKeyStatus(AIPortAccessKey accessKey);
+    int updateAccessKeyStatus(AIPortToolAccessKey accessKey);
 
     @Delete("DELETE FROM "+accessKeyTable+" WHERE user_Id=#{userId} and id=#{id}")
     int deleteAccessKey(@Param("userId")long userId,@Param("id") long id);